Adiciona Controller e lógica das transferencias

+++ transferencia/src/main/java/com/agendamento/transferencia/model/Transferencia.java
@@ -35,5 +35,5 @@ public class Transferencia {
     private LocalDate dataTransferencia;
 
     @Column(nullable = false)
-    private BigDecimal taxa;
+    private Double taxa;
 }
